refactor: remove header & align content
