+++ web/beacon-app/src/features/projects/components/APIKeysTable.tsx
@@ -1,6 +1,6 @@
 import { t, Trans } from '@lingui/macro';
-import { Button, Heading, Table, Toast } from '@rotational/beacon-core';
-import { useEffect, useState } from 'react';
+import { Button, Heading, Loader, Table, Toast } from '@rotational/beacon-core';
+import { Suspense, useEffect, useState } from 'react';
 
 import { ApiKeyModal } from '@/components/common/Modal/ApiKeyModal';
 import { HelpTooltip } from '@/components/common/Tooltip/HelpTooltip';
@@ -97,69 +97,80 @@ export const APIKeysTable = ({ projectID }: APIKeysTableProps) => {
   ];
 
   return (
-    <div className="mt-[46px]  border-y-neutral-600" data-cy="keyComp">
-      <Heading as={'h1'} className="flex items-center text-lg font-semibold capitalize">
-        <Trans>Permission Your Data Flows: Generate API Keys</Trans>
-      </Heading>
-      <div className="flex space-x-1">
-        <p className="my-4">
-          <Trans>
-            Now that you have at least one topic or event stream set up, generate API keys to set
-            permissions and securely provision data access. API keys connect your publishers (data
-            sources) to Ensign and control access by subscribers. Generate at least one API key for
-            your project. You can customize permissions.
-          </Trans>
-          <span className="ml-2" data-cy="keyHint">
-            <HelpTooltip data-cy="keyInfo">
-              <p>
-                <Trans>
-                  Each key consists of two parts - a ClientID and a ClientSecret. You'll need both
-                  to establish a client connection, create Ensign topics, publishers, and
-                  subscribers. Keep your API keys private -- if you misplace your keys, you can
-                  revoke them and generate new ones.
-                </Trans>
-              </p>
-            </HelpTooltip>
-          </span>
-        </p>
-      </div>
-      <div className="flex w-full justify-between bg-[#F7F9FB] p-2">
-        <div className="flex items-center gap-3"></div>
-        <div>
-          <Button
-            variant="primary"
-            size="small"
-            className="px-5 !text-xs"
-            onClick={onOpenGenerateAPIKeyModal}
-            data-cy="addKey"
-          >
-            + New Key
-          </Button>
+    <Suspense
+      fallback={
+        <div className="flex justify-center">
+          <Loader />
         </div>
-      </div>
-      <Table
-        trClassName="text-sm"
-        columns={initialColumns}
-        data={getApiKeys(apiKeys, {
-          handleOpenRevokeAPIKeyModal,
-        })}
-        data-cy="keyTable"
-      />
-      {openRevokeAPIKeyModal.opened && (
-        <RevokeAPIKeyModal onOpen={openRevokeAPIKeyModal} onClose={handleCloseRevokeAPIKeyModal} />
-      )}
-      {isOpenAPIKeyDataModal && (
-        <ApiKeyModal open={isOpenAPIKeyDataModal} data={key} onClose={onCloseAPIKeyDataModal} />
-      )}
-      {isOpenGenerateAPIKeyModal && (
-        <GenerateAPIKeyModal
-          open={isOpenGenerateAPIKeyModal}
-          onClose={onCloseGenerateAPIKeyModal}
-          onSetKey={setKey}
-          projectId={projectID}
+      }
+    >
+      <div className="mt-[46px]  border-y-neutral-600" data-cy="keyComp">
+        <Heading as={'h1'} className="flex items-center text-lg font-semibold capitalize">
+          <Trans>Permission Your Data Flows: Generate API Keys</Trans>
+        </Heading>
+        <div className="flex space-x-1">
+          <p className="my-4">
+            <Trans>
+              Now that you have at least one topic or event stream set up, generate API keys to set
+              permissions and securely provision data access. API keys connect your publishers (data
+              sources) to Ensign and control access by subscribers. Generate at least one API key
+              for your project. You can customize permissions.
+            </Trans>
+            <span className="ml-2" data-cy="keyHint">
+              <HelpTooltip data-cy="keyInfo">
+                <p>
+                  <Trans>
+                    Each key consists of two parts - a ClientID and a ClientSecret. You'll need both
+                    to establish a client connection, create Ensign topics, publishers, and
+                    subscribers. Keep your API keys private -- if you misplace your keys, you can
+                    revoke them and generate new ones.
+                  </Trans>
+                </p>
+              </HelpTooltip>
+            </span>
+          </p>
+        </div>
+        <div className="flex w-full justify-between bg-[#F7F9FB] p-2">
+          <div className="flex items-center gap-3"></div>
+          <div>
+            <Button
+              variant="primary"
+              size="small"
+              className="px-5 !text-xs"
+              onClick={onOpenGenerateAPIKeyModal}
+              data-cy="addKey"
+            >
+              + New Key
+            </Button>
+          </div>
+        </div>
+        <Table
+          trClassName="text-sm"
+          columns={initialColumns}
+          data={getApiKeys(apiKeys, {
+            handleOpenRevokeAPIKeyModal,
+          })}
+          data-cy="keyTable"
         />
-      )}
-    </div>
+        {openRevokeAPIKeyModal.opened && (
+          <RevokeAPIKeyModal
+            onOpen={openRevokeAPIKeyModal}
+            onClose={handleCloseRevokeAPIKeyModal}
+          />
+        )}
+        {isOpenAPIKeyDataModal && (
+          <ApiKeyModal open={isOpenAPIKeyDataModal} data={key} onClose={onCloseAPIKeyDataModal} />
+        )}
+        {isOpenGenerateAPIKeyModal && (
+          <GenerateAPIKeyModal
+            open={isOpenGenerateAPIKeyModal}
+            onClose={onCloseGenerateAPIKeyModal}
+            onSetKey={setKey}
+            projectId={projectID}
+          />
+        )}
+      </div>
+    </Suspense>
   );
 };
