+++ web/beacon-app/src/features/projects/components/ProjectDetail/DetailInfo.tsx
@@ -1,7 +1,5 @@
 import { Trans } from '@lingui/macro';
-import { Heading } from '@rotational/beacon-core';
 
-import { TagState } from '@/components/common/TagState';
 import { SentryErrorBoundary } from '@/components/Error';
 import { formatDate } from '@/utils/formatDate';
 
@@ -11,7 +9,7 @@ interface ProjectDetailInfoProps {
 }
 
 const ProjectDetailInfo: React.FC<ProjectDetailInfoProps> = ({ data }) => {
-  const { description, status, created, owner } = data || {};
+  const { description, created, owner } = data || {};
 
   const getFormattedDecription = () => {
     if (!description) {
@@ -32,20 +30,9 @@ const ProjectDetailInfo: React.FC<ProjectDetailInfoProps> = ({ data }) => {
         </div>
       }
     >
-      <div className="mt-[46px]">
-        <Heading as={'h1'} className=" flex items-center text-lg font-semibold capitalize">
-          <Trans>Project Details</Trans>
-        </Heading>
+      <div className="my-[16px] border-b border-gray-500 ">
         <table className="mt-4 table-auto border-separate border-spacing-y-4">
           <tbody>
-            <tr>
-              <td className="w-[150px] font-semibold">
-                <Trans>Project Status:</Trans>
-              </td>
-              <td>
-                <TagState status={status as string} />
-              </td>
-            </tr>
             <tr>
               <td className="w-[150px] font-semibold">
                 <Trans>Description:</Trans>
