+++ web/beacon-app/src/features/projects/components/TopicTable.tsx
@@ -1,6 +1,6 @@
 import { t, Trans } from '@lingui/macro';
 import { Button, Loader, Table } from '@rotational/beacon-core';
-import { useEffect, useMemo, useState } from 'react';
+import { Suspense, useEffect, useMemo, useState } from 'react';
 import toast from 'react-hot-toast';
 import { useNavigate, useParams } from 'react-router-dom';
 
@@ -41,40 +41,48 @@ export const TopicTable = () => {
   }, [error, hasTopicsFailed]);
 
   return (
-    <div className="mt-[46px]" data-cy="topicComp">
-      {isFetchingTopics ? (
-        <Loader />
-      ) : (
-        <>
-          <TopicTableHeader />
-          <div className="flex w-full justify-between bg-[#F7F9FB] p-2">
-            <div className="flex items-center gap-3"></div>
-            <Button
-              variant="primary"
-              size="small"
-              className="!text-xs"
-              onClick={handleOpenNewTopicModal}
-              data-cy="addTopic"
-            >
-              <Trans>+ New Topic</Trans>
-            </Button>
-            <NewTopicModal open={openNewTopicModal} handleClose={handleCloseNewTopicModal} />
-          </div>
-          <div className="overflow-hidden text-sm">
-            <Table
-              trClassName="text-sm"
-              columns={initialColumns}
-              data={getTopics(topics)}
-              onRowClick={(row: any) => {
-                redirectToTopicDetails(row?.values?.id);
-              }}
-              initialState={getHiddenColumns(['id'])}
-              data-cy="topicTable"
-            />
-          </div>
-        </>
-      )}
-    </div>
+    <Suspense
+      fallback={
+        <div className="flex justify-center">
+          <Loader />
+        </div>
+      }
+    >
+      <div className="mt-[46px]" data-cy="topicComp">
+        {isFetchingTopics ? (
+          <Loader />
+        ) : (
+          <>
+            <TopicTableHeader />
+            <div className="flex w-full justify-between bg-[#F7F9FB] p-2">
+              <div className="flex items-center gap-3"></div>
+              <Button
+                variant="primary"
+                size="small"
+                className="!text-xs"
+                onClick={handleOpenNewTopicModal}
+                data-cy="addTopic"
+              >
+                <Trans>+ New Topic</Trans>
+              </Button>
+              <NewTopicModal open={openNewTopicModal} handleClose={handleCloseNewTopicModal} />
+            </div>
+            <div className="overflow-hidden text-sm">
+              <Table
+                trClassName="text-sm"
+                columns={initialColumns}
+                data={getTopics(topics)}
+                onRowClick={(row: any) => {
+                  redirectToTopicDetails(row?.values?.id);
+                }}
+                initialState={getHiddenColumns(['id'])}
+                data-cy="topicTable"
+              />
+            </div>
+          </>
+        )}
+      </div>
+    </Suspense>
   );
 };
