+++ web/beacon-app/src/features/projects/routes/ProjectDetailPage.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import { Heading, Loader } from '@rotational/beacon-core';
+
 import invariant from 'invariant';
-import { lazy, Suspense, useEffect } from 'react';
+import { lazy, useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 
 import { PATH_DASHBOARD } from '@/application/routes/paths';
@@ -10,6 +10,7 @@ import AppLayout from '@/components/layout/AppLayout';
 import ProjectActive from '../components/ProjectActive';
 import ProjectBreadcrumbs from '../components/ProjectBreadcrumbs';
 import ProjectDetailInfo from '../components/ProjectDetail/DetailInfo';
+import ProjectDetailHeader from '../components/ProjectDetail/ProjectDetailHeader';
 import ProjectSetup from '../components/ProjectSetup';
 import ProjectSettings from '../components/Settings';
 import { useFetchProject } from '../hooks/useFetchProject';
@@ -28,10 +29,6 @@ const ProjectDetailPage = () => {
     useProjectSetup(projectID);
   const { project, error } = useFetchProject(projectID);
 
-  const getNormalizedProjectName = () => {
-    return project?.name?.split('-').join(' ');
-  };
-
   useEffect(() => {
     // when user switch to another organization and the current project is not found then redirect to projects page
     if (error?.response?.status === 401) {
@@ -42,46 +39,27 @@ const ProjectDetailPage = () => {
   return (
     <AppLayout Breadcrumbs={<ProjectBreadcrumbs project={project} />}>
       <div className="flex items-center justify-between rounded-md bg-[#F7F9FB] px-6 py-3">
-        <Heading as="h1" className="flex items-center text-lg font-semibold">
-          <span className="mr-2" data-cy="project-name">
-            {getNormalizedProjectName()}
-          </span>
-        </Heading>
+        <ProjectDetailHeader data={project} />
         <ProjectSettings data={project} />
       </div>
-      <ProjectDetailInfo data={project} />
-      {!hasAlreadySetup && (
-        <ProjectSetup
-          warningMessage={warningMessage}
-          config={{
-            isProjectCreated: hasProject,
-            isAPIKeyCreated: hasApiKeys,
-            isTopicCreated: hasTopics,
-          }}
-        />
-      )}
-      {!isActive && hasAlreadySetup && (
-        <ProjectActive onActive={setIsActive} projectID={projectID} />
-      )}
-
-      <Suspense
-        fallback={
-          <div className="flex justify-center">
-            <Loader />
-          </div>
-        }
-      >
+      <div className="mx-6">
+        <ProjectDetailInfo data={project} />
+        {!hasAlreadySetup && (
+          <ProjectSetup
+            warningMessage={warningMessage}
+            config={{
+              isProjectCreated: hasProject,
+              isAPIKeyCreated: hasApiKeys,
+              isTopicCreated: hasTopics,
+            }}
+          />
+        )}
+        {!isActive && hasAlreadySetup && (
+          <ProjectActive onActive={setIsActive} projectID={projectID} />
+        )}
         <TopicTable />
-      </Suspense>
-      <Suspense
-        fallback={
-          <div className="flex justify-center">
-            <Loader />
-          </div>
-        }
-      >
         <APIKeysTable projectID={projectID} />
-      </Suspense>
+      </div>
     </AppLayout>
   );
 };
